data_processing/requirements.txt
yfinance==0.2.55
numpy>=1.16.5
pandas
requests
confluent-kafka
pyspark
flask
psycopg2-binary